backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chemschema"
version = "0.1.0"
description = "ChemSchema – a reactive chemistry structure editor Django app with gRPC and REST API"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
authors = [
    { name = "LARA Chemistry", email = "info@lara-chemistry.example.com" },
]
keywords = [
    "chemistry",
    "django",
    "editor",
    "molecules",
    "reactions",
    "smiles",
    "molfile",
    "grpc",
    "rdkit",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Chemistry",
]
requires-python = ">=3.10"
dependencies = [
    "django>=4.2",
    "django-ninja>=1.6",
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "protobuf>=4.25",
    "django-cors-headers>=4.3",
]

[project.optional-dependencies]
rdkit = [
    "rdkit>=2023.9",
]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=5.0",
    "black>=24.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "django-stubs>=5.0",
]

[project.urls]
Homepage = "https://github.com/LARA-chemistry/chemschema"
Repository = "https://github.com/LARA-chemistry/chemschema.git"
Issues = "https://github.com/LARA-chemistry/chemschema/issues"
Documentation = "https://github.com/LARA-chemistry/chemschema#readme"

# ── Hatch build configuration ──────────────────────────────────────────────────

[tool.hatch.build.targets.sdist]
include = [
    "chemschema/",
    "proto/",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["chemschema"]

# ── pytest ────────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
pythonpath = ["."]
testpaths = ["tests"]
addopts = "--tb=short -q"

# ── ruff ──────────────────────────────────────────────────────────────────────

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"**/migrations/*.py" = ["E501", "N806"]

# ── black ────────────────────────────────────────────────────────────────────

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

# ── mypy ─────────────────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "tests.settings"
